requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
requests==2.31.0
plotly==5.19.0
python-dotenv==1.0.0
streamlit-aggrid==0.3.4.post3